backend/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
openai
pydantic
python-dotenv
celery
redis
asyncpg
